+++ lib/generators/webflow_sync/templates/webflow_sync.rb
@@ -5,5 +5,4 @@ WebflowSync.configure do |config|
   # config.skip_webflow_sync = Rails.env.test? # default
   config.webflow_site_id = ENV.fetch('WEBFLOW_SITE_ID')
   config.sync_webflow_slug = ENV.fetch('SYNC_WEBFLOW_SLUG')
-  config.publish_on_sync = true
 end