+++ spec/jobs/webflow_sync/update_item_job_spec.rb
@@ -13,12 +13,10 @@ module WebflowSync
     let(:webflow_site_id) { ENV.fetch('WEBFLOW_SITE_ID') }
 
     before(:each) do
-      @old_publish_on_sync = WebflowSync.configuration.publish_on_sync
       @old_sync_webflow_slug = WebflowSync.configuration.sync_webflow_slug
       @old_webflow_site_id = WebflowSync.configuration.webflow_site_id
 
       WebflowSync.configure do |config|
-        config.publish_on_sync = false
         config.sync_webflow_slug = sync_webflow_slug
         config.webflow_site_id = webflow_site_id
       end
@@ -26,7 +24,6 @@ module WebflowSync
 
     after(:each) do
       WebflowSync.configure do |config|
-        config.publish_on_sync = @old_publish_on_sync
         config.sync_webflow_slug = @old_sync_webflow_slug
         config.webflow_site_id = @old_webflow_site_id
       end