+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 6.0.0
+
+- As long as the error is `Rate limit hit`, sleep 10 seconds and retry the request
+- Lower the Rails version requirement to >= 5.0
+- Remove `WebFlow.configuration.publish_on_sync` config option. This is no longer needed. We rely on Webflow correctly updating collection items by sending `live='true'` parameter on every create/update/delete request.
+- Call `WebflowSync::UpdateItemJob` in `WebflowSync::CreateItemJob` if record already contains `webflow_item_id`. This can sometimes happen when we create a record in Rails, then update the record, so now there are `WebflowSync::CreateItemJob` and `WebflowSync::UpdateItemJob` jobs for the record, and then for some reason, first `WebflowSync::CreateItemJob` job fails. (This can happen for example because `Rate limit hit` or some other Webflow API error.) `WebflowSync::UpdateItemJob` will then run before `WebflowSync::CreateItemJob` and it will call `WebflowSync::CreateItemJob`. After that, when the original `WebflowSync::CreateItemJob` runs, it will create another record and overwrite the `webflow_item_id` created by `WebflowSync::UpdateItemJob`.
+- This gem currently only works with a fork of 'webflow-ruby': `gem 'webflow-ruby', github: 'vfonic/webflow-ruby', branch: 'allow-live-delete'`
+
+# 5.0.0
+
+- Require Rails >= 7.0
+
 # 4.0.2
 
 - Revert "Explicitly require Webflow::Client and Webflow::Error" (v4.0.2 is the same as v4.0.0)